Creating wrapper types for passing in values
This is to improve the separation of the JSInterpreter and jsed. 
So now invoke has a more generic interface, that takes a vector of js::ValueRefs as parameters, rather than a concrete signature. 
Reasonable types are automatically converted into ValueRefs.

This also prepares the move to GCC 4.7, which supports some interesting features that come
with c++11.